orchestrator/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
anthropic==0.39.0
openai==1.58.1
GitPython==3.1.43
requests==2.32.3
apscheduler==3.10.4
python-dotenv==1.0.1
